requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
tensorboard>=2.8.0
pyyaml>=5.4.0
transformers>=4.30.0
einops>=0.6.0
wandb>=0.12.0
jupyter>=1.0.0
pytest>=7.0.0